+++ src/server/socket-class/index.js
@@ -1,0 +1,3 @@
+import SocketClass from './socket-class';
+
+export default SocketClass;